fix Qt build on Ubuntu amd64

+++ src/orm/qtsql_driver.cpp
@@ -89,7 +89,7 @@ QtSqlCursorBackend::fetch_row()
                 v = Value(stmt_->value(i).toLongLong());
             else if (t == QVariant::DateTime || t == QVariant::Date)
                 v = Value(stmt_->value(i).toDateTime());
-            else if (t == QVariant::Double || t == QMetaType::Float)
+            else if (t == QVariant::Double || (int)t == (int)QMetaType::Float)
                 v = Value(stmt_->value(i).toDouble());
             else
                 v = Value(stmt_->value(i).toString());
